combine service and sensor listener
Now that sensors need reregistering on the fly, sensor listener needs to
hold onto the service for its Context. This makes room for potential
inconsistency where service can be null for the listener or vice versa.

Combine these classes since they do the same thing.
